updating to refactor
