+++ scripts/azure-price-fetch/README.md
@@ -3,5 +3,5 @@
 ## How to test
 
 ```sh
-> AZURE_SUBSCRIPTION_ID="<sub ID>" go run ./scripts/azure-price-fetch/main.go
+> AZURE_SUBSCRIPTION_ID="<sub ID>" go run ./scripts/azure-price-fetch/.
 ```
